segundaParte/proyectoTraductorLSA/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pandas==2.2.2
spacy==3.7.4
moviepy==1.0.3
python-multipart==0.0.9
openai-whisper==20231117
torch==2.2.1

